--- book/projects/jnibookgradle/build.gradle
@@ -1,23 +1,0 @@
-plugins {
-    // Apply the java plugin to add support for Java
-    id 'java'
-
-    // Apply the application plugin to add support for building a CLI application.
-    id 'application'
-}
-
-repositories {
-    // Use jcenter for resolving dependencies.
-    // You can declare any Maven/Ivy/file repository here.
-    jcenter()
-}
-
-dependencies {
-    // Use JUnit test framework
-    testImplementation 'junit:junit:4.13'
-}
-
-application {
-    // Define the main class for the application.
-    mainClassName = 'com.github.jni_rs.jnibook.App'
-}
