--- book/projects/jnibookgradle/src/main/java/com/github/jni_rs/jnibook/App.java
@@ -1,7 +1,0 @@
-package com.github.jni_rs.jnibook;
-
-public class App {
-    public static void main(String[] args) {
-        NativeAPI.verify_link();
-    }
-}
