--- book/projects/jnibookgradle/src/test/java/com/github/jni_rs/jnibook/NativeAPITest.java
@@ -1,13 +1,0 @@
-package com.github.jni_rs.jnibook;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-public class NativeAPITest {
-
-    @Test
-    public void verifyLink() {
-        NativeAPI.verifyLink();
-    }
-}
