--- book/projects/jnibookrs/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "jnibookrs"
-version = "0.1.0"
-authors = [""]
-edition = "2018"
-publish = false
-
-[dependencies]
-jni = "0.17"
-lazy_static = "1.4"
-
-[lib]
-crate_type = ["cdylib"]
-
-[profile.release]
-debug = true
-lto = true
