+++ book/projects/starter/jnibookgradle/.gitattributes
