+++ book/projects/starter/jnibookgradle/.gitignore
