+++ book/projects/starter/jnibookgradle/build.gradle
@@ -1,0 +1,30 @@
+plugins {
+    // Apply the java plugin to add support for Java
+    id 'java'
+
+    // Apply the application plugin to add support for building a CLI application.
+    id 'application'
+}
+
+repositories {
+    // Use jcenter for resolving dependencies.
+    // You can declare any Maven/Ivy/file repository here.
+    jcenter()
+}
+
+dependencies {
+    // Use JUnit test framework
+    testImplementation 'junit:junit:4.13'
+}
+
+application {
+    // Define the main class for the application.
+    mainClassName = 'jni_rs_book.App'
+    applicationDefaultJvmArgs = ['']
+}
+
+tasks.withType(Test) {
+    systemProperties System.properties
+}
+
+test.outputs.upToDateWhen { false }
