+++ book/projects/starter/jnibookgradle/gradle/wrapper/gradle-wrapper.properties
