+++ book/projects/starter/jnibookgradle/settings.gradle
