+++ book/projects/starter/jnibookgradle/src/main/java/jni_rs_book/App.java
@@ -1,0 +1,7 @@
+package jni_rs_book;
+
+public class App {
+    public static void main(String[] args) {
+        NativeAPI.verifyLink();
+    }
+}
