+++ book/projects/starter/jnibookgradle/src/test/java/jni_rs_book/NativeAPITest.java
@@ -1,0 +1,11 @@
+package jni_rs_book;
+
+import org.junit.Test;
+
+public class NativeAPITest {
+
+    @Test
+    public void verifyLink() {
+        NativeAPI.verifyLink();
+    }
+}
