+++ book/projects/starter/jnibookrs/src/counter.rs
@@ -1,0 +1,1 @@
+// Empty on purpose
