+++ book/projects/starter/jnibookrs/src/lib.rs
@@ -1,0 +1,4 @@
+pub mod counter;
+pub mod division;
+pub mod charts;
+pub mod verifylink;
