+++ book/projects/starter/jnibookrs/src/verifylink.rs
@@ -1,0 +1,5 @@
+use jni::objects::JClass;
+use jni::JNIEnv;
+
+#[no_mangle]
+pub extern "system" fn Java_jni_1rs_1book_NativeAPI_verify_1link(_env: JNIEnv, _class: JClass) {}
