+++ book/src/SUMMARY.md
@@ -1,10 +1,18 @@
 # Summary
 
 - [Introduction and Environment Setup](./introduction.md)
-- [Linking and Dividing](./linking_and_dividing.md)
-- [Debugging](./debugging.md)
-- [Panics and Results](./error_handling.md)
-- [Building a Counter](./counter.md)
-- [Cleaning Up Resource Leaks](./cleaning.md)
-- [Java Objects in Rust](./java_objects_in_rust.md)
+- [Linking](./linking.md)
+- [Division Exercise](./division_exercise.md)
+  - [Division Hint](./division_hint.md)
+  - [Division Solution](./division_solution.md)
+- [Error Handling](./error_handling.md)
+- [Chart Exercise](./chart_exercise.md)
+  - [Chart Hint](./chart_hint.md)
+  - [Chart Solution](./chart_solution.md)
+- [Counter Exercise](./counter_exercise.md)
+  - [Counter Hint](./counter_hint.md)
+  - [Counter Solution](./counter_solution.md)
+  - [Counter Discussion](./counter_discussion.md)
+- [Future Topics](./future_topics.md)
 - [Appendix](./appendix.md)
+  - [Debugging](./debugging.md)
