+++ book/src/division_summary.md
@@ -1,0 +1,10 @@
+# Division Exercise Summary
+
+To recap, in the division exercise, you:
+
+1. Wrote JNI code for Rust to perform division.
+2. May have tested panics across the JNI boundary, and seen a crash (undefined
+   behavior).
+
+In the next chapter, we'll explore a few error handling patterns so that we are
+less likely to observe jvm crashes.
